requirements.txt
numpy==1.19.4
opencv-python==4.4.0.46
opencv-contrib-python==4.4.0.46
python-dotenv==0.15.0